% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollateData.R
\name{CollateData}
\alias{CollateData}
\title{Processes data from IRFinder output}
\usage{
CollateData(Experiment, reference_path, output_path)
}
\arguments{
\item{Experiment:}{A data frame containing the necessary information. The first column is the base name of the sample files. The second column is the sample name. All subsequent columns are assumed to be experimental conditions / annotation data (e.g. gender, age, treatment response, etc)}

\item{reference_path:}{THe path to the reference generated by BuildReference()}

\item{output_path:}{The path to the output generated by data collation.}
}
\description{
Processes data from IRFinder output
}
