% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_filter}
\alias{default_filter}
\title{Default / recommended filters for NxtIRF}
\format{
A list of 8 filters:
\itemize{
\item Depth filter (requires total splice depth of > 20 (splice + IR depth for IR events))
\item Coverage filter IR (90\% coverage for IR events)
\item Coverage filter AS (binary isoforms tested must represent > 60\% of total events across given region)
\item Consistency filter MXE / SE (for tandem splice events, each event must be at least 1/4 of total)
}
}
\usage{
default_filter
}
\description{
This is default filters object. The variables are as follows:
}
\keyword{datasets}
