% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{download_NxtIRF_example}
\alias{download_NxtIRF_example}
\title{Downloads NxtIRF example BAM files for the vignette}
\usage{
download_NxtIRF_example(destination_dir = tempdir())
}
\arguments{
\item{destination_dir}{The path to store the downloaded BAM files}
}
\description{
Downloads NxtIRF example BAM files for the vignette
}
