% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{FetchAnnotation}
\alias{FetchAnnotation}
\title{Fetch genome / transcriptome reference from AnnotationHub and writes to reference}
\usage{
FetchAnnotation(
  ah_genome = "AH65745",
  ah_transcriptome = "AH64631",
  reference_path = "./Reference",
  verbose = FALSE
)
}
\description{
Fetch genome / transcriptome reference from AnnotationHub and writes to reference
}
