% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{GenerateMappability}
\alias{GenerateMappability}
\title{Fetch genome / transcriptome reference from AnnotationHub and writes to reference}
\usage{
GenerateMappability(
  fasta = "genome.fa",
  gtf = "transcripts.gtf",
  ah_transcriptome = "",
  reference_path = "./Reference",
  read_len = 70,
  read_stride = 10,
  error_pos = 35,
  verbose = FALSE
)
}
\description{
Fetch genome / transcriptome reference from AnnotationHub and writes to reference
}
