% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{GenerateMappabilityReads}
\alias{GenerateMappabilityReads}
\title{Fetch genome / transcriptome reference from AnnotationHub and writes to reference_path}
\usage{
GenerateMappabilityReads(
  fasta = "genome.fa",
  gtf = "transcripts.gtf",
  ah_genome = "",
  ah_transcriptome = "",
  reference_path = "./Reference",
  read_len = 70,
  read_stride = 10,
  error_pos = 35,
  verbose = FALSE,
  localHub = FALSE
)
}
\description{
Fetch genome / transcriptome reference from AnnotationHub and writes to reference_path
}
