% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{NxtAnnotation}
\alias{NxtAnnotation}
\title{The NxtAnnotation Class}
\description{
The class that stores NxtIRF reference
}
\section{Fields}{

\describe{
\item{\code{chrOrder}}{Chromosome names}

\item{\code{Genes}}{Gene annotations (GRanges)}

\item{\code{Trascripts}}{Transcript annotations (GRanges)}

\item{\code{Exons}}{Exon annotations (GRanges)}

\item{\code{Introns}}{Introns (data.table). Contains duplicate introns for each transcripts that share the same genomic interval. For unique introns, use Introns_dedup}

\item{\code{Introns_dedup}}{Introns, unique for each genomic interval (data.table).}

\item{\code{Splice}}{Splice annotations (data.table)}
}}

