% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{NxtIRF.BuildReference}
\alias{NxtIRF.BuildReference}
\title{Builds NxtIRF Reference}
\usage{
NxtIRF.BuildReference(
  fasta = "genome.fa",
  gtf = "transcripts.gtf",
  RefPath = "",
  stopcodons = c("TAA", "TAG", "TGA"),
  PolyA_length = 15,
  PolyA_mismatch = 1
)
}
\arguments{
\item{fasta:}{Either the path to the genome fasta file, or the name of the AnnotationHub object containing the TwoBitFile of the genome (beginning with "AH...")}

\item{gtf:}{Either the path to the transcript gtf file, or the name of the AnnotationHub object containing the GRanges object of the transcript reference (beginning with "AH...")}

\item{RefPath:}{The directory in which genome.fa and transcripts.gtf is found. If not provided, a temp directory will be generated.}

\item{stopcodons:}{A vector containing stop codons for your species of interest. Default is c("TAA","TAG","TGA")}

\item{PolyA_length:}{(Default = 15) Minimum length of repeat adenosine sequence detection.}

\item{PolyA_mismatch:}{(Default = 1) Maximum number of allowed mismatch of repeat adenosine sequence detection.}
}
\value{
Nothing. The created NxtAnnotation object is saved in 'RefPath/Ref_Filename'
}
\description{
Builds reference files used by NxtIRF. Requires Ensembl GTF file 'transcripts.gtf' and genome file 'genome.fa'
}
