% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{nxtIRF}
\alias{nxtIRF}
\title{Launches the NxtIRF Graphics User Interface using Shiny Dashboard}
\usage{
nxtIRF(offline = FALSE)
}
\arguments{
\item{offline}{Whether to run in offline mode (i.e. local mode for 
AnnotationHub). Set as `FALSE` to run online mode}
}
\value{
None
}
\description{
This function launches the NxtIRF interactive app using Shiny Dashboard
}
