% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Classes.R
\docType{class}
\name{NxtProject}
\alias{NxtProject}
\title{The NxtProject Class}
\description{
The main class used by NxtIRF to store project information.
}
\section{Fields}{

\describe{
\item{\code{metadata}}{list, contains Project, Data, and Bam directory paths, as well as given name of data Header}

\item{\code{metrics}}{list, contains metrics that describe the nature of the data (stranded vs unstranded) as well as various parameters from which the raw IRFinder data is compiled. See NxtIRF.ProcessData()}

\item{\code{Expr}}{list, contains list of sample annotations. The entire data set is retrieved as Expr$global or Expr[["global"]]. Subset analysis is also possible by sub-setting the Expr$global data frame.}

\item{\code{Data}}{list, contains compiled raw data loaded into memory.}

\item{\code{nds}}{list, contains refined countData with accompanying colData, for parsing onto other package methods for differential analysis such as DESeq2, edgeR or limma}

\item{\code{benchmarks}}{list, contains system times for benchmark purposes}
}}

