\name{rIRFinder-package}
\alias{rIRFinder-package}
\alias{rIRFinder}
\docType{package}
\title{
\packageTitle{rIRFinder}
}
\description{
\packageDescription{rIRFinder}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rIRFinder}
\packageIndices{rIRFinder}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rIRFinder}

Maintainer: \packageMaintainer{rIRFinder}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
