% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{theme_white}
\alias{theme_white}
\title{A theme object for white background figures without a legend}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 10.
}
\usage{
theme_white
}
\description{
A theme object for white background figures without a legend
}
\seealso{
[NxtIRF::theme_white_legend]
}
\keyword{datasets}
