% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{theme_white_legend}
\alias{theme_white_legend}
\title{A theme object for white background figures with a legend}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 9.
}
\usage{
theme_white_legend
}
\description{
A theme object for white background figures with a legend
}
\seealso{
[NxtIRF::theme_white]
}
\keyword{datasets}
