% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{BuildReference}
\alias{BuildReference}
\title{Builds IRFinder Reference}
\usage{
BuildReference(
  fasta = "genome.fa",
  gtf = "transcripts.gtf",
  ah_genome = "",
  ah_transcriptome = "",
  reference_path = "./Reference",
  genome_type = "",
  nonPolyARef = "",
  MappabilityRef = "",
  BlacklistRef = "",
  FilterIRByProcessedTranscript = FALSE
)
}
\arguments{
\item{fasta:}{The path to the genome fasta file}

\item{gtf:}{The path to the transcript gtf file}

\item{reference_path:}{The directory to write the (r)/IRFinder reference files}
}
\value{
Nothing. The created reference will be written to the given directory
}
\description{
Builds reference files used by IRFinder / rIRFinder. Requires genome fasta and transcriptome gtf file
}
