% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{FetchAnnotation}
\alias{FetchAnnotation}
\title{Fetch genome / transcriptome reference from AnnotationHub and writes to reference}
\usage{
FetchAnnotation(
  ah_genome = "",
  ah_transcriptome = "",
  reference_path = "./Reference",
  verbose = FALSE
)
}
\description{
E.g. hg38 release-94: genome = "AH65745", transcriptome = "AH64631"
}
