% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{nxtIRF}
\alias{nxtIRF}
\title{Launches the NxtIRF Graphics User Interface using Shiny Dashboard}
\usage{
nxtIRF(offline = FALSE, BPPARAM = BiocParallel::bpparam())
}
\description{
Launches the NxtIRF Graphics User Interface using Shiny Dashboard
}
